% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\docType{data}
\name{cps}
\alias{cps}
\title{cps dataset}
\format{
A very very large dataset
\describe{
\item{C_SOP}{Scores for self-oriented perfectionism}
\item{C_SSP}{Scores for socially prescribed perfectionism}
\item{C_OOP}{Scores for other-oriented perfectionism}
}
}
\usage{
cps
}
\value{
Perfectionism and Psychopathology Lab
}
\description{
A dataset containing perfectionism scores
}
\keyword{datasets}
