% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL_J.R
\name{FL_J}
\alias{FL_J}
\title{Filtering Loss for a set of filtered taxa J}
\usage{
FL_J(X, J)
}
\arguments{
\item{X}{OTU table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}

\item{J}{A vector of J taxa to be removed. It must be subset of column names of X.}
}
\value{
FL Filtering loss value
}
\description{
This function calculates filtering loss due to removing a group of J taxa.
}
\examples{
data(mock2)

# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts

#arrange counts in order of increasing number of samples taxa are present in
NP <- NP_Order(Counts)
Counts <- Counts[,NP]

# Extract the taxa names to be removed
J <- colnames(Counts)[1:30]

#Calculate filtering loss due to removing these taxa
FL_J(Counts,J)

}
\references{
Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutation  filtration of microbiome data", to be submitted.
}
\seealso{
\code{\link{FiltLoss}}
}
\author{
Ekaterina Smirnova
}
