% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{knight}
\alias{knight}
\title{Bacterial Diversity in Neonatal Intensive Care Units (NICUs) data}
\format{This file contains....}
\usage{
data(knight)
}
\description{
These data (Knights et al., 2011) with 30 samples and 1097 taxa was collected to investigat the sources of bacteria
 found on surfaces and equipment in NICU. The data taxa was previously analyzed using sourcetracker software to identify
 the proportion of bacteria from each environment using published datasets from environments likely to be sources of indoor
 contaminants, namely human skin, oral cavities, feces and temperate soils.
}
\keyword{datasets}
