% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{mock}
\alias{mock}
\title{Oral artificial communities data}
\format{This phyloseq object contains an OTU table with 1016 samples and 27211 taxa, its metadata and taxomomy.}
\usage{
data(mock)
}
\description{
This dataset comes from the MBQC project, a collaborative effort designed to comprehensively evaluate sample processing
 and computational methods for human microbiome data analysis (Sinha et al., 2015). There are four types of samples in
 this project: (1) 11 unique fresh stool samples; (2) seven unique freeze-dried stool samples; (3) two unique chemostat
 samples generated from a Robogut; and (4) two artificial colonies representing the gut and oral cavity. These samples
 were randomly sequenced at 15 laboratories and then randomly distributed to 9 bioinformatics facilities for microbiome
 analyses. Here, we consider the oral artificial communities data which comprised of 22 true taxa from the human
 oral cavity.
}
\keyword{datasets}
