% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{mock2}
\alias{mock2}
\title{Bias experiment data}
\format{This file contains a count OTU table and a proportion OTU table, each with 240 samples and 46 taxa. A list of true
 taxa is also given.}
\usage{
data(mock2)
}
\description{
These publicly available data (Brooks et al., 2015) were generated as a part of a study designed to evaluate the bias
 at each step of the VCU sequencing protocol, namely,  DNA extraction,  PCR amplification,  sequencing and taxonomic
 classification. Mock community samples were created out of 7 vaginally relevant bacteria  by mixing prescribed
 quantities of cells, with quantities varying across samples according to an experimental design described in
 Brooks et al, 2015. As opposed to the positive controls data, bacteria appear in different proportions across samples.
 The number of taxa identified by the sequencing and bioinformatics pipeline was 46.
}
\keyword{datasets}
