% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{NCw_Order}
\alias{NCw_Order}
\title{Taxa importance ordering by the weighted number of connected taxa}
\usage{
NCw_Order(Counts)
}
\arguments{
\item{Counts}{OTU COUNTS table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}
}
\value{
NCW Taxa names in increasing  order of the weighted number of connected taxa.
}
\description{
Taxa importance ordering by the weighted number of connected taxa
}
\examples{
data(mock2)
# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts

#arrange counts in order of increasing number of samples taxa are present in
NCw <- NCw_Order(Counts)

}
\references{
Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutation  filtration of microbiome data", to be submitted.
}
\author{
Ekaterina Smirnova
}
