% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{NP_Order}
\alias{NP_Order}
\title{Taxa importance ordering by the number of occurrences of the taxa in the n samples}
\usage{
NP_Order(Counts)
}
\arguments{
\item{Counts}{OTU COUNTS table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}
}
\value{
NP Taxa names in increasing  order of the number of samples taxa are present in.
}
\description{
Taxa importance ordering by the number of occurrences of the taxa in the n samples
}
\examples{
data(mock2)
# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts

#arrange counts in order of increasing number of samples taxa are present in
NP <- NP_Order(Counts)
}
\references{
Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutation  filtration of microbiome data", to be submitted.
}
\author{
Ekaterina Smirnova
}
