\name{PERFect-package}
\alias{PERFect-package}
\alias{PERFect}
\docType{package}
\title{
\packageTitle{PERFect}
}
\description{
\packageDescription{PERFect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PERFect}
\packageIndices{PERFect}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PERFect}

Maintainer: \packageMaintainer{PERFect}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
