% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pvals_Order}
\alias{pvals_Order}
\title{Taxa importance ordering by PERFect p-values}
\usage{
pvals_Order(Counts, res_sim)
}
\arguments{
\item{Counts}{OTU COUNTS table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}

\item{res_sim}{Output of \code{PERFect_sim()} function.}
}
\value{
Order_pvals Taxa names in increasing order of p-values significance.
}
\description{
This function orders taxa by increasing significance of simultaneous PERFect p-values.
}
\examples{
data(mock2)

# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts

# Perform simultaenous filtering of the data
res_sim <- PERFect_sim(X=Counts)

#order according to p-values
pvals_sim <- pvals_Order(Counts, res_sim)

}
\references{
Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutation  filtration of microbiome data", to be submitted.
}
\seealso{
\code{\link{PERFect_sim}}, \code{\link{PERFect_perm}}
}
\author{
Ekaterina Smirnova
}
