% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals_Plots.R
\name{pvals_Plots}
\alias{pvals_Plots}
\title{Plots of PERFect p-values}
\usage{
pvals_Plots(PERFect, X, quantiles = c(0.25, 0.5, 0.8, 0.9), alpha = 0.05)
}
\arguments{
\item{PERFect}{Output of \code{PERFect_sim()} or \code{PERFect_perm()} function.}

\item{X}{OTU table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}

\item{quantiles}{Quantile values for coloring, these are set to 25\%, 50\%, 80\% and
90\% percentiles of the individual filtering loss values.}

\item{alpha}{Alpha level of the test, set to 0.1 by default.}
}
\value{
p_vals Plot of p-values
}
\description{
Graphical representation of p-values obtained by running \code{PERFect_sim()} or
\code{PERFect_perm()} for jth taxon colored by quantile values of individual filtering loss.
}
\examples{
data(mock2)
# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts
dim(Counts) # 240x46

# Perform simultaenous filtering of the data
res_sim <- PERFect_sim(X=Counts)
dim(res_sim$filtX)      # 240x10, removing 36 taxa
colnames(res_sim$filtX) # signal taxa

#permutation perfect colored by FLu values
pvals_Plots(PERFect = res_sim, X = Counts, quantiles = c(0.25, 0.5, 0.8, 0.9), alpha=0.05)
}
\seealso{
\code{\link{PERFect_sim}}, \code{\link{PERFect_perm}}
}
\author{
Ekaterina Smirnova
}
