% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{ravel}
\alias{ravel}
\title{Ravel vaginal microbiome data}
\format{This file contains the OTU tables in counts and proportion, as well as its metadata. The OTU table has 394 samples
 and 247 taxa.}
\usage{
data(ravel)
}
\description{
The vaginal microbiome study of a cohort of 396 reproductive age women previously published in (Ravel et al., 2011),
 where data and details on data collection and preprocessing are also available. The goal of the study was to understand
 the role and ultimate function of vaginal microbiota in reducing risk of infectious diseases and to identify factors
 leading to disease susceptibility. Microbiome data were obtained by pyrosequencing of barcoded 16S rRNA genes;
 in this data set 247 taxa were identified.
}
\keyword{datasets}
