% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{salter}
\alias{salter}
\title{Reagent and laboratory contamination data}
\format{This file contains the OTU table, its metadata and taxomomy. The OTU table has 42 samples and 625 taxa.}
\usage{
data(salter)
}
\description{
The data (Salter et al., 2014) was generated from the study of the effect of present contaminants in DNA extraction kits
 and other laboratory reagents on sequencing DNA. Mock samples of a pure Salmonella bongori culture had undergone five
 rounds of serial ten-fold dilutions to generate a series of high to low biomass samples. To generate a taxa counts table
 from this study, we used samples for the Salmonella bongori culture 16S rRNA gene profiling data, which are deposited as
 FASTQ files under ENA project accession EMBL: ERP006737 (https://www.ebi.ac.uk/ena/data/view/PRJEB7055),
 and processed using the dada2 R-package.
}
\keyword{datasets}
