% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TraditR1}
\alias{TraditR1}
\title{Traditional Filtering Rule 1}
\usage{
TraditR1(X, thresh =5)
}
\arguments{
\item{X}{OTU COUNTS table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}

\item{thresh}{Numerical value, set to 5 by default. Throughout all samples, taxa that are present for less than this
threshold with be removed.}
}
\value{
\item{filtX}{Filtered OTU table}
}
\description{
This rule suggests to remove taxa that are mostly absent in all samples.
}
\examples{
data(mock2)

# Counts data matrix
Counts <- mock2$Counts

# Filtering
Filtered_X <- TraditR1(Counts)

}
\references{
Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutation  filtration of microbiome data.
}
\author{
Ekaterina Smirnova
}
