% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TraditR2}
\alias{TraditR2}
\title{Traditional Filtering Rule 2}
\usage{
TraditR2(X, Ab_min = 0.001)
}
\arguments{
\item{X}{OTU COUNTS table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}

\item{Ab_min}{Numerical value, set to 0.001 by default. Throughout all samples, taxa with abundance less than this
threshold with be removed.}
}
\value{
\item{filtX}{Filtered OTU table}
}
\description{
This rule is adopted from Milici et al. (2016) that removes taxa with low abundance level.
 Specifically, it keeps taxa with abundance level higher than 0.001%.
 Then it further selects taxa that satisfy at least one of the following conditions:  Present in at least one sample
 at a relative abundance higher than 1% of the reads of that sample, present in at least 2% of samples at a relative
 abundance higher than 0.1% for a given sample, present in at least 5% of samples at any abundance level.
}
\examples{
data(mock2)

# Counts data matrix
Counts <- mock2$Counts

# Filtering
Filtered_X <- TraditR2(Counts)

}
\references{
Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutation  filtration of microbiome data.
}
\author{
Ekaterina Smirnova
}
