% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERFect_perm.R
\name{PERFect_perm}
\alias{PERFect_perm}
\title{Permutation PERFect filtering for microbiome data}
\usage{
PERFect_perm(X, infocol = NULL, Order = "NP", Order.user = NULL, normalize = "counts",
    algorithm = "fast", center = FALSE, quant = c(0.1, 0.25, 0.5),
    distr = "sn", alpha = 0.1, lag = 3, direction = "left", pvals_sim = NULL,
    k = 10000, nbins = 30, hist = TRUE, col = "red", fill = "green",
    hist_fill = 0.2, linecol = "blue")
}
\arguments{
\item{X}{OTU table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.}

\item{infocol}{Index vector of the metadata. We assume user only gives a taxa table,
but if the metadata of the samples are included in the columns of the input, this option
needs to be specified.}

\item{Order}{Taxa ordering. The default ordering is the number of occurrences (NP) of the taxa in all samples.
Other types of order are p-value ordering, number of connected taxa and weighted number of connected taxa,
denoted as \code{"pvals"}, \code{"NC"}, \code{"NCw"} respectively. More details about taxa ordering are described in Smirnova et al.
User can also specify their preference order with Order.user.}

\item{Order.user}{User's taxa ordering. This argument takes a character vector of ordered taxa names.}

\item{normalize}{Normalizing taxa count. The default option does not normalize taxa count,
but user can convert the OTU table into a proportion table using the option \code{"prop"}
or convert it into a presence/absence table using \code{"pres"}.}

\item{algorithm}{Algorithm speed. The default is speed is \code{"fast"}, which allows the program to efficiently
search for significant taxa without computing all the p-values. User must use the default option \code{"hist = FALSE"}
for the fast algorithm. The alternative setting is \code{"full"}, which computes all the taxa's p-values.}

\item{center}{Centering OTU table. The default option does not center the OTU table.}

\item{quant}{Quantile values used to fit the distribution to log DFL values.
The number of quantile values corresponds to the number of parameters in the distribution the data is fitted to.
Assuming that at least 50\% of taxa  are not informative, we suggest fitting the log Skew-Normal distribution
by matching the 10\%, 25\% and 50\% percentiles of the log-transformed samples to the Skew-Normal distribution.}

\item{distr}{The type of distribution to fit log DFL values to. While we suggest using Skew-Normal distribution,
and set as the default distribution, other choices are available.
\describe{
\item{\code{"sn"}}{Skew-Normal distribution with 3 parameters: location xi, scale omega^2 and shape alpha}
\item{\code{"norm"}}{Normal distribution with 2 parameters: mean and standard deviation sd}
}}

\item{alpha}{Test level alpha, set to 0.1 by default.}

\item{lag}{Integer width of the rolling window in rolling average (moving mean), set to 3 by default.}

\item{direction}{Character specifying whether the index of the result should be left- or right-aligned
or centered compared to the rolling window of observations, set to "left" by default.}

\item{pvals_sim}{Object resulting from simultaneous PERFect with taxa abundance ordering,
allowing user to perform Simultaneous PERFect with p-values ordering.
Be aware that the choice of distribution for both methods must be the same.}

\item{k}{The number of permutations, set to 10000 by default.}

\item{nbins}{Number of bins used to visualize the histogram  of log DFL values, set to 30 by default.}

\item{hist}{Binary TRUE/FALSE value. If TRUE, the function builds histograms for each taxon.}

\item{col}{Graphical parameter for  color of histogram bars border, set to "red" by default.}

\item{fill}{Graphical parameter for color of histogram fill, set to "green" by default.}

\item{hist_fill}{Graphical parameter for intensity of histogram fill, set to 0.2 by default.}

\item{linecol}{Graphical parameter for the color of the fitted distribution density, set to "blue" by default.}
}
\value{
\item{filtX}{Filtered OTU table}

\item{pvals}{P-values of the test}

\item{DFL}{Differences in filtering loss values}

\item{fit}{Fitted values and further goodness of fit details passed from the \code{fitdistr()} function}

\item{hist}{Histogram of log differences in filtering loss}

\item{est}{Estimated distribution parameters}

\item{dfl_distr}{Plot of differences in filtering loss values}
}
\description{
Permutation filtering of the provided OTU table X at a test level alpha.
  Each set of j taxa significance is evaluated by fitting the Skew-Normal, Normal, t or Cauchy distribution
  to the  sampling distribution obtained by permuted taxa labels.
}
\details{
Filtering is the process of identifying and removing a subset of taxa according to a particular criterion.
  As opposed to the the simultaneous filtering approach, we do not assume that all distributions
  for each set of taxa  are identical and equal to the distribution of simultaneous filtering.
  Function PERFect_perm()  filters  the provided OTU table X and outputs a filtered table that contains signal taxa.
  PERFect_perm() calculates differences in filtering loss DFL for each taxon according to the given taxa order.
  By default, the function fits Skew-Normal distribution to the log-differences in filtering loss but Normal,
  t, or Cauchy distributions can be also used.
}
\examples{
data(mock2)

# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts

# Perform simultaenous filtering of the data
res_sim <- PERFect_sim(X=Counts)

#order according to p-values
pvals_sim <- pvals_Order(Counts, res_sim)

#### Uncomment to run algorithm with parallel processing ith more than 2 cores
# #obtain permutation PERFEct results using NP taxa ordering
# res_perm <- PERFect_perm(X = Prop, Order.user = pvals_sim, algorithm = "fast")

# #permutation perfect colored by FLu values
# pvals_Plots(PERFect = res_perm, X = Counts, quantiles = c(0.25, 0.5, 0.8, 0.9), alpha=0.05)
}
\references{
Azzalini, A. (2005). The skew-normal distribution and related multivariate families. Scandinavian Journal of Statistics, 32(2), 159-188.

Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutationfiltration of microbiome data", to be submitted.
}
\seealso{
\code{\link{PERFect_sim}}
}
\author{
Ekaterina Smirnova
}
