% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PERFect_sim.R
\name{PERFect_sim}
\alias{PERFect_sim}
\title{Simulation PERFect filtering for microbiome data}
\usage{
PERFect_sim(X,infocol = NULL, Order = "NP", Order.user = NULL, normalize = "counts",
         center = FALSE, quant = c(0.1, 0.25, 0.5), distr = "sn",
         alpha = 0.1, lag = 3, direction = "left", pvals_sim = NULL,
         nbins = 30, col = "red", fill = "green", hist_fill = 0.2,
         linecol = "blue")
}
\arguments{
\item{X}{OTU table, where taxa are columns and samples are rows of the table.
It should be a in data frame format with columns corresponding to taxa names.
It could contains columns of metadata.}

\item{infocol}{Index vector of the metadata. We assume user only gives a taxa table,
but if the metadata of the samples are included in the columns of the input, this option
needs to be specified.}

\item{Order}{Taxa ordering. The default ordering is the number of occurrences (NP) of the taxa in all samples.
Other types of order are p-value ordering, number of connected taxa and weighted number of connected taxa,
denoted as \code{"pvals"}, \code{"NC"}, \code{"NCw"} respectively. More details about taxa ordering are described in Smirnova et al.
User can also specify their preference order with Order.user.}

\item{normalize}{Normalizing taxa count. The default option does not normalize taxa count,
but user can convert the OTU table into a proportion table using the option \code{"prop"}
or convert it into a presence/absence table using \code{"pres"}.}

\item{center}{Centering OTU table. The default option does not center the OTU table.}

\item{quant}{Quantile values used to fit the distribution to log DFL values.
The number of quantile values corresponds to the number of parameters in the distribution the data is fitted to.
Assuming that at least 50\% of taxa  are not informative, we suggest fitting the log Skew-Normal distribution
by matching the 10\%, 25\% and 50\% percentiles of the log-transformed samples to the Skew-Normal distribution.}

\item{distr}{The type of distribution to fit log DFL values to. While we suggest using Skew-Normal distribution,
and set as the default distribution, other choices are available.
\describe{
\item{\code{"sn"}}{Skew-Normal distribution with 3 parameters: location xi, scale omega^2 and shape alpha}
\item{\code{"norm"}}{Normal distribution with 2 parameters: mean and standard deviation sd}
\item{\code{"t"}}{Student t-distribution with 2 parameters: n degrees of freedom and noncentrality ncp}
\item{\code{"cauchy"}}{Cauchy distribution with 2 parameters: location and scale}
}}

\item{alpha}{Test level alpha, set to 0.1 by default.}

\item{lag}{Integer width of the rolling window in rolling average (moving mean), set to 3 by default.}

\item{direction}{Character specifying whether the index of the result should be left- or right-aligned
or centered compared to the rolling window of observations, set to "left" by default.}

\item{pvals_sim}{Object resulting from simultaneous PERFect with taxa abundance ordering,
allowing user to perform Simultaneous PERFect with p-values ordering.
Be aware that the choice of distribution for both methods must be the same.}

\item{nbins}{Number of bins used to visualize the histogram  of log DFL values, set to 30 by default.}

\item{col}{Graphical parameter for  color of histogram bars border, set to "red" by default.}

\item{fill}{Graphical parameter for color of histogram fill, set to "green" by default.}

\item{hist_fill}{Graphical parameter for intensity of histogram fill, set to 0.2 by default.}

\item{linecol}{Graphical parameter for the color of the fitted distribution density, set to "blue" by default.}
}
\value{
\item{filtX}{Filtered OTU table}

\item{pvals}{P-values of the test}

\item{DFL}{Differences in filtering loss values}

\item{fit}{Fitted values and further goodness of fit details passed from the \code{fitdistr()} function}

\item{hist}{Histogram of log differences in filtering loss}

\item{est}{Estimated distribution parameters}

\item{pDFL}{Plot of differences in filtering loss values}
}
\description{
Simultaneous filtering of the provided OTU table X at a test level alpha. One distribution is fit to taxa simultaneously.
}
\details{
Filtering is the process of identifying and removing a subset of taxa according to a particular criterion.
Function \code{PERFect_sim()}  filters  the provided OTU table X and outputs a filtered table
that contains signal taxa. \code{PERFect_sim()} calculates differences in filtering loss DFL
for each taxon according to the given taxa order. By default, the function fits Skew-Normal distribution
to the log-differences in filtering loss but Normal, t, or Cauchy distributions can be also used.
This is implementation of Algorithm 1 described in Smirnova et al.
}
\examples{
data(mock2)
# Proportion data matrix
Prop <- mock2$Prop

# Counts data matrix
Counts <- mock2$Counts
dim(Counts) # 240x46

# Perform simultaenous filtering of the data
res_sim <- PERFect_sim(X=Counts)
dim(res_sim$filtX)      # 240x10, removing 36 taxa
colnames(res_sim$filtX) # signal taxa

#permutation perfect colored by FLu values
pvals_Plots(PERFect = res_sim, X = Counts, quantiles = c(0.25, 0.5, 0.8, 0.9), alpha=0.05)

}
\references{
Azzalini, A. (2005). The skew-normal distribution and related multivariate families. Scandinavian Journal of Statistics, 32(2), 159-188.

Smirnova, E., Huzurbazar, H., Jafari, F. ``PERFect: permutationfiltration of microbiome data", to be submitted.
}
\seealso{
\code{\link{PERFect_perm}}
}
\author{
Ekaterina Smirnova
}
